fix: renamed useWindow to useWindowSize
