+++ packages/react/src/index.ts
@@ -2,4 +2,4 @@ export { default as useBooleanState } from './hooks/useBooleanState';
 export { default as usePreviousState } from './hooks/usePreviousState';
 export { default as useThrottle } from './hooks/useThrottle';
 export { default as useDebounce } from './hooks/useDebounce';
-export { default as useWindow } from './hooks/useWindow';
+export { default as useWindow } from './hooks/useWindowSize';
